moved models, reverted vars in google model
